finished

+++ src/main/java/com/comeon/demo/entity/CountedEmail.java
@@ -6,10 +6,10 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import javax.persistence.*;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 import java.sql.Timestamp;
-import java.util.HashSet;
-import java.util.Set;
 
 @Entity
 @Builder
@@ -28,5 +28,5 @@ public class CountedEmail {
 
     private String email;
 
-    private int count;
+    private long count;
 }
