+++ src/main/java/com/comeon/demo/model/Email.java
@@ -1,6 +1,5 @@
 package com.comeon.demo.model;
 
-import com.fasterxml.jackson.databind.jsonschema.JsonSerializableSchema;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -13,7 +12,10 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @AllArgsConstructor
 @NoArgsConstructor
+
 public class Email {
+
     private String email;
-    private int count;
+
+    private long count;
 }
